Fix lint errors

+++ common/log/base_logger.go
@@ -45,12 +45,12 @@ const unixDateMilli = "Mon Jan _2 15:04:05.000 MST 2006"
 // LogConfig : Configuration to be provided to logging infra
 type LogFileConfig struct {
 	LogFile      string
-	LogSize      uint64
-	LogFileCount int
+	LogSize      int64
+	LogFileCount int64
 	LogLevel     common.LogLevel
 	LogTag       string
 
-	currentLogSize uint64
+	currentLogSize int64
 }
 
 type BaseLogger struct {
@@ -135,18 +135,18 @@ func (l *BaseLogger) SetLogFile(name string) error {
 			l.logFileHandle = f
 			fi, e := f.Stat()
 			if e == nil {
-				l.fileConfig.currentLogSize = uint64(fi.Size())
+				l.fileConfig.currentLogSize = int64(fi.Size())
 			}
 		}
 	}
 	return nil
 }
 
-func (l *BaseLogger) SetMaxLogSize(size int) {
-	l.fileConfig.LogSize = uint64(size) * 1024 * 1024
+func (l *BaseLogger) SetMaxLogSize(size int64) {
+	l.fileConfig.LogSize = int64(size) * 1024 * 1024
 }
 
-func (l *BaseLogger) SetLogFileCount(count int) {
+func (l *BaseLogger) SetLogFileCount(count int64) {
 	l.fileConfig.LogFileCount = count
 }
 
@@ -180,7 +180,7 @@ func (l *BaseLogger) init() error {
 	} else {
 		fi, e := os.Stat(l.fileConfig.LogFile)
 		if e == nil {
-			l.fileConfig.currentLogSize = uint64(fi.Size())
+			l.fileConfig.currentLogSize = fi.Size()
 		}
 		var err error
 		l.logFileHandle, err = os.OpenFile(l.fileConfig.LogFile, os.O_CREATE|os.O_WRONLY|os.O_APPEND, 0644)
@@ -234,7 +234,7 @@ func (l *BaseLogger) logDumper(id int, channel <-chan string) {
 	for j := range channel {
 		l.logger.Println(j)
 
-		l.fileConfig.currentLogSize += (uint64)(len(j))
+		l.fileConfig.currentLogSize += (int64)(len(j))
 		if l.fileConfig.currentLogSize > l.fileConfig.LogSize {
 			//fmt.Println("Calling logrotate : ", l.fileConfig.currentLogSize, " : ", l.fileConfig.logSize)
 			_ = l.LogRotate()
